+++ app/Models/Receiving.php
@@ -24,13 +24,17 @@ class Receiving extends Model
 
             $old_amount = $query->getOriginal('amount');
             $new_amount = $query->amount;
-
-            // fetch invoice
-            $invoice = Invoice::find($query->invoice_id);
+            $old_invoice_id = $query->getOriginal('invoice_id');
+            $new_invoice_id = $query->invoice_id;
 
             // old
-            // update invoice
-            $invoice->amount_pay -= $old_amount;
+            // if($old_invoice_id != NULL){
+            //     // fetch invoice
+            //     $invoice = Invoice::find($old_invoice_id);
+            //     // update invoice
+            //     $invoice->amount_pay -= $old_amount;
+            //     $invoice->save();
+            // }
 
             // ledger entry
             // Ledger::create([
@@ -39,33 +43,40 @@ class Receiving extends Model
             //     'type' => 'debit',
             //     'transaction_date' => return_todays_date()
             // ]);
-            $ledger = Ledger::where('customer_id', $invoice->customer_id)
+            $ledger = Ledger::where('customer_id', $query->customer_id)
                             ->where('receiving_id', $query->id)
                             ->where('amount', $old_amount)
                             ->first();
-            $ledger->delete();
+            if($ledger){
+                $ledger->delete();
+            }
 
             // new
-            // update invoice
-            $invoice->amount_pay += $new_amount;
+            // if($new_invoice_id != NULL){
+            //     // fetch invoice
+            //     $invoice = Invoice::find($new_invoice_id);
+            //     // update invoice
+            //     $invoice->amount_pay += $new_amount;
+            //     $invoice->save();
+            // }
 
             // ledger entry
             Ledger::create([
-                'customer_id' => $invoice->customer_id,
+                'customer_id' => $query->customer_id,
                 'receiving_id' => $query->id,
                 'amount' => $new_amount,
                 'type' => 'credit',
                 'transaction_date' => return_todays_date()
             ]);
-
-            $invoice->save();
         });
 
         static::deleting(function ($query) {
-            // update invoice
-            $invoice = Invoice::find($query->invoice_id);
-            $invoice->amount_pay -= $query->amount;
-            $invoice->save();
+            // if($query->invoice_id != NULL){
+            //     // update invoice
+            //     $invoice = Invoice::find($query->invoice_id);
+            //     $invoice->amount_pay -= $query->amount;
+            //     $invoice->save();
+            // }
 
             // ledger entry
             // Ledger::create([
@@ -74,22 +85,27 @@ class Receiving extends Model
             //     'type' => 'debit',
             //     'transaction_date' => return_todays_date()
             // ]);
-            $ledger = Ledger::where('customer_id', $invoice->customer_id)
+            $ledger = Ledger::where('customer_id', $query->customer_id)
                             ->where('receiving_id', $query->id)
                             ->where('amount', $query->amount)
                             ->first();
-            $ledger->delete();
+            if($ledger){
+                $ledger->delete();
+            }
         });
 
         static::created(function ($query) {
-            // update invoice
-            $invoice = Invoice::find($query->invoice_id);
-            $invoice->amount_pay += $query->amount;
-            $invoice->save();
+            // if($query->invoice_id != NULL){
+            //     // update invoice
+            //     $invoice = Invoice::find($query->invoice_id);
+            //     $invoice->amount_pay += $query->amount;
+            //     $invoice->save();
+            // }
 
             // ledger entry
             Ledger::create([
-                'customer_id' => $invoice->customer_id,
+                'customer_id' => $query->customer_id,
+                'invoice_id' => $query->invoice_id ? $query->invoice_id : NULL,
                 'receiving_id' => $query->id,
                 'amount' => $query->amount,
                 'type' => 'credit',
@@ -99,7 +115,7 @@ class Receiving extends Model
     }
     
     protected $fillable = [
-        'invoice_id', 'amount', 'created_by', 'modified_by',
+        'invoice_id', 'customer_id', 'amount', 'created_by', 'modified_by',
     ];
 
     protected $dates = ['created_at', 'updated_at', 'deleted_at'];
@@ -112,5 +128,10 @@ class Receiving extends Model
     public function ledgers()
     {
         return $this->hasMany('App\Models\Ledger');
+    }
+
+    public function customer()
+    {
+        return $this->belongsTo('App\Models\Customer');
     }
 }
